refactor: applicate naming rule
