+++ src/pages/todos/todo-detail/containers/todo-detail.container.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import TodoDetail from "../components/TodoDetail";
+import TodoDetail from "../components/todo-detail.component";
 
 const TodoDetailContainer: React.FC<{ id: number }> = (props) => {
   return <TodoDetail id={1} />;
