+++ src/pages/todos/todo.route.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import TodoDetailContainer from "./todo-detail/containers/TodoDetailContainer";
+import TodoDetailContainer from "./todo-detail/containers/todo-detail.container";
 
 const Routes: React.FC = () => {
   return (
